requirements.txt
Django==1.8.17
django-mock-queries==0.0.16.5
factory-boy==2.8.1
